refactor : List, Grid 중복코드 통합
List, Gird의 ContentView의 중복된 코드를 프로토콜로 통합
- Contentable
