+++ OpenMarket/OpenMarket/View/ItemList/ItemListContentView.xib
@@ -8,11 +8,11 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="ItemContentView" customModule="OpenMarket" customModuleProvider="target">
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="ItemListContentView" customModule="OpenMarket" customModuleProvider="target">
             <connections>
-                <outlet property="ItemImageView" destination="gbs-ZO-5JF" id="Vhq-cx-1n9"/>
                 <outlet property="containerView" destination="iN0-l3-epB" id="RyU-f1-P3v"/>
                 <outlet property="discountPrice" destination="BHM-pa-8eN" id="sg8-aw-b0x"/>
+                <outlet property="itemImageView" destination="gbs-ZO-5JF" id="Vhq-cx-1n9"/>
                 <outlet property="price" destination="Msg-cl-Hz2" id="Zin-Lf-QCT"/>
                 <outlet property="stock" destination="NTx-Mf-M3z" id="MdX-mh-tQx"/>
                 <outlet property="title" destination="ZB8-Q6-pxW" id="Ixl-Lk-gIH"/>
